feat: move whiteboard component to a page + remove whiteboard-index component
