+++ src/app/whiteboard/[id]/page.css
